+++ src/styles/globalStyle.ts
