+++ src/views/Index.tsx
@@ -1,13 +1,18 @@
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import NavigationBar from './layouts/NavigationBar';
+import NavigationBar from '../components/NavigationBar';
 import Home from './pages/Home';
 import NotFound from './pages/NotFound';
+import ChatLists from './pages/ChatLists';
 import React from 'react';
 const Index = () => {
   const routeLists = [
     {
       path: '/',
       element: <Home />,
+    },
+    {
+      path: '/chatLists',
+      element: <ChatLists />,
     },
     {
       path: '/reviews',
