+++ src/views/pages/NotFound.tsx
@@ -1,4 +1,4 @@
-import '../../../style/NotFound.css';
+import '../../styles/NotFound.css';
 import React from 'react';
 function NotFound() {
   return (
